relocate func
