+++ src/server/handlers/manager/join.rs
@@ -3,9 +3,9 @@
 //! Join subnet handler and parameters
 
 use crate::manager::SubnetManager;
-use crate::server::handlers::manager::check_subnet;
 use crate::server::handlers::manager::subnet::SubnetManagerPool;
-use crate::server::{parse_from, JsonRPCRequestHandler};
+use crate::server::handlers::manager::{check_subnet, parse_from};
+use crate::server::JsonRPCRequestHandler;
 use anyhow::anyhow;
 use async_trait::async_trait;
 use fvm_shared::econ::TokenAmount;
