+++ src/server/handlers/manager/leave.rs
@@ -3,9 +3,9 @@
 //! Leave subnet handler and parameters
 
 use crate::manager::SubnetManager;
-use crate::server::handlers::manager::check_subnet;
 use crate::server::handlers::manager::subnet::SubnetManagerPool;
-use crate::server::{parse_from, JsonRPCRequestHandler};
+use crate::server::handlers::manager::{check_subnet, parse_from};
+use crate::server::JsonRPCRequestHandler;
 use anyhow::anyhow;
 use async_trait::async_trait;
 use ipc_sdk::subnet_id::SubnetID;
