+++ src/server/handlers/manager/mod.rs
@@ -9,6 +9,8 @@ pub mod subnet;
 
 use crate::config::Subnet;
 use anyhow::{anyhow, Result};
+use fvm_shared::address::Address;
+use std::str::FromStr;
 
 pub(crate) fn check_subnet(subnet: &Subnet) -> Result<()> {
     if subnet.auth_token.is_none() {
@@ -16,4 +18,19 @@ pub(crate) fn check_subnet(subnet: &Subnet) -> Result<()> {
         return Err(anyhow!("Internal server error"));
     }
     Ok(())
+}
+
+pub(crate) fn parse_from(subnet: &Subnet, from: Option<String>) -> Result<Address> {
+    let addr = match from {
+        Some(addr) => Address::from_str(&addr)?,
+        None => {
+            if subnet.accounts.is_empty() {
+                log::error!("subnet does not have account defined, {:?}", subnet.id);
+                return Err(anyhow!("Internal server error"));
+            } else {
+                subnet.accounts[0]
+            }
+        }
+    };
+    Ok(addr)
 }
